samples/bot-task-modules/python/bot-task-modules/pyproject.toml
[project]
name = "bot-task-modules"
version = "0.1.0"
description = "This sample demonstrates Task Modules in Microsoft Teams using the Teams SDK for Python."
readme = "README.md"
requires-python = ">=3.12,<3.15"
dependencies = [
    "python-dotenv>=1.0.*",
    "microsoft-teams-apps==2.0.0a9",
]